Move fib to trait tests
