+++ crypto/openstark/examples/large_fib.rs
@@ -2,7 +2,7 @@
 use env_logger;
 use log::info;
 use macros_decl::field_element;
-use openstark::{Constraints, TraceTable, RationalExpression, proof, verify, Provable, Verifiable};
+use openstark::{proof, verify, Constraints, Provable, RationalExpression, TraceTable, Verifiable};
 use primefield::FieldElement;
 use std::{env, time::Instant};
 use u256::U256;
